refactor(terms): Extract term functionality
