+++ src/ducks/apps/index.js
@@ -34,6 +34,7 @@ import {
   fetchAppOrKonnector
 } from './client-helpers'
 import { APP_TYPE, APP_STATE, REGISTRY_CHANNELS } from './constants'
+import termUtils from './terms'
 import flatten from 'lodash/flatten'
 
 export { APP_STATE }
@@ -42,7 +43,6 @@ export { REGISTRY_CHANNELS }
 
 const APPS_DOCTYPE = 'io.cozy.apps'
 const KONNECTORS_DOCTYPE = 'io.cozy.konnectors'
-const TERMS_DOCTYPE = 'io.cozy.terms'
 
 const DEFAULT_CHANNEL = storeConfig.default.registry.channel
 
@@ -493,56 +493,6 @@ export function uninstallApp(app) {
   }
 }
 
-let termsIndexCache = null
-async function _getOrCreateTermsIndex() {
-  termsIndexCache = await cozy.client.data.defineIndex(TERMS_DOCTYPE, [
-    'termsId',
-    'version'
-  ])
-  return termsIndexCache
-}
-
-async function _saveAppTerms(terms) {
-  const { id, ...termsAttributes } = terms
-  // We use : as separator for <id>:<version>
-  let savedTermsDocs = null
-  try {
-    savedTermsDocs = await cozy.client.data.query(
-      await _getOrCreateTermsIndex(),
-      {
-        selector: {
-          termsId: id,
-          version: termsAttributes.version
-        },
-        limit: 1
-      }
-    )
-  } catch (e) {
-    throw e
-  }
-  if (savedTermsDocs && savedTermsDocs.length) {
-    // we just update the url if this is the same id and same version
-    // but the url changed
-    const savedTerms = savedTermsDocs[0]
-    if (
-      savedTerms.termsId == id &&
-      savedTerms.version == termsAttributes.version &&
-      savedTerms.url != termsAttributes.url
-    ) {
-      await cozy.client.data.updateAttributes(TERMS_DOCTYPE, savedTerms._id, {
-        url: termsAttributes.url
-      })
-    }
-  } else {
-    const termsToSave = Object.assign({}, termsAttributes, {
-      termsId: id,
-      accepted: true,
-      acceptedAt: new Date()
-    })
-    await cozy.client.data.create(TERMS_DOCTYPE, termsToSave)
-  }
-}
-
 export function installApp(client, app, source, isUpdate = false) {
   const { slug, type, terms } = app
   return async dispatch => {
@@ -559,7 +509,7 @@ export function installApp(client, app, source, isUpdate = false) {
       .join('&')
     if (terms) {
       try {
-        await _saveAppTerms(terms)
+        await termUtils.save(terms)
       } catch (e) {
         handleError(e)
       }
